changed amount column to be a float and migrated

+++ db/migrate/20140718144728_add_amount_to_posts.rb
@@ -1,5 +1,5 @@
 class AddAmountToPosts < ActiveRecord::Migration
   def change
-    add_column :posts, :amount, :integer
+    add_column :posts, :amount, :float
   end
 end
